requirements.txt
Django==3.2
flake8==3.9.0
djangorestframework==3.12.4